% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{equation_values}
\alias{equation_values}
\title{Extract equation records}
\usage{
equation_values(result, equation)
}
\arguments{
\item{result}{A result created by \code{solve()}.}

\item{equation}{An equation object or equation name.}
}
\value{
A data frame of equation records.
}
\description{
Extract equation records
}
