% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gams.R
\name{extract_status_from_lst}
\alias{extract_status_from_lst}
\title{Extract solver and model status from a lst file}
\usage{
extract_status_from_lst(lstfile)
}
\arguments{
\item{lstfile}{path to the lst file to parse}
}
\value{
A 'tibble()' with two columns: solver_status and model_status for all
solve summary.
}
\description{
Extract solver and model status from a lst file
}
