% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{format_gams_expression}
\alias{format_gams_expression}
\title{Format a symbolic expression as GAMS code}
\usage{
format_gams_expression(expression)
}
\arguments{
\item{expression}{A GAMSr symbolic expression.}
}
\value{
A single character string containing deterministic GAMS expression
code.
}
\description{
Format a symbolic expression as GAMS code
}
