% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{gams_eq}
\alias{gams_eq}
\title{Create a symbolic equality relation}
\usage{
gams_eq(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{GAMSr symbolic expressions or numeric scalar constants.}
}
\value{
A \code{gams_expr_comparison} object.
}
\description{
\code{gams_eq()} creates a symbolic equality relation for equation definitions.
It maps to the GAMS \verb{=e=} equation relationship during rendering.
}
