% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation.R
\name{gams_equation}
\alias{gams_equation}
\title{Create a GAMS equation}
\usage{
gams_equation(model, name, domain = NULL, description = NULL)
}
\arguments{
\item{model}{A model created by \code{\link[=gams_model]{gams_model()}}.}

\item{name}{GAMS identifier for the equation.}

\item{domain}{Optional set or set vector created with \code{c(i, j)}.}

\item{description}{Optional explanatory text.}
}
\value{
An immutable \code{gams_equation} symbol object registered with \code{model}.
}
\description{
Create a GAMS equation
}
