% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{gams_math}
\alias{gams_math}
\alias{gams_abs}
\alias{gams_exp}
\alias{gams_log}
\alias{gams_log10}
\alias{gams_sqrt}
\alias{gams_min}
\alias{gams_max}
\title{GAMSr mathematical functions}
\usage{
gams_abs(x)

gams_exp(x)

gams_log(x)

gams_log10(x)

gams_sqrt(x)

gams_min(...)

gams_max(...)
}
\arguments{
\item{x}{A GAMSr symbolic expression or numeric scalar.}

\item{...}{GAMSr symbolic expressions or numeric scalars.}
}
\value{
A \code{gams_expr_math_function} object.
}
\description{
These helpers create symbolic math-function nodes. They do not evaluate their
arguments as ordinary R vectors.
}
