% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{gams_same_as}
\alias{gams_same_as}
\title{Compare set elements or labels}
\usage{
gams_same_as(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{GAMSr sets, aliases, or single character labels. At least one
operand must be a set or alias.}
}
\value{
A symbolic condition rendered with GAMS \code{sameAs()}.
}
\description{
Compare set elements or labels
}
