% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{gams_transfer_available}
\alias{gams_transfer_available}
\title{Check for GAMS Transfer R}
\usage{
gams_transfer_available()
}
\value{
\code{TRUE} when the optional \code{gamstransfer} package can be loaded.
}
\description{
Check for GAMS Transfer R
}
