% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{gamstransfer_adapter}
\alias{gamstransfer_adapter}
\title{Create a GAMS Transfer R adapter}
\usage{
gamstransfer_adapter(system_directory = NULL)
}
\arguments{
\item{system_directory}{Optional GAMS system directory passed to
\code{gamstransfer::Container$new()}.}
}
\value{
A \code{gams_transfer_adapter} object.
}
\description{
Create a GAMS Transfer R adapter
}
