% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{generated_gams}
\alias{generated_gams}
\title{Generate deterministic GAMS source}
\usage{
generated_gams(problem)
}
\arguments{
\item{problem}{A problem created by \code{\link[=gams_problem]{gams_problem()}}.}
}
\value{
A single character string containing GAMS source code.
}
\description{
Generate deterministic GAMS source
}
