% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gdxfile.R
\name{import_gdxfile}
\alias{import_gdxfile}
\title{Function to import all the sets, scalars, parameters and variables from a gdx file}
\usage{
import_gdxfile(filename, selection = NULL)
}
\arguments{
\item{filename}{Name of the gdxfile to be imported.}

\item{selection}{Character vector with information on which sets, scalars, parameters or variables to read}
}
\value{
All sets, scalars, parameters, and variables in the gdxfile
}
\description{
Function to import all the sets, scalars, parameters and variables from a gdx file
}
\examples{
\dontrun{import_gdxfile("results.gdx")}
}
