% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_gamstransfer.R
\name{install_gamstransfer}
\alias{install_gamstransfer}
\title{Install gamstransfer package}
\usage{
install_gamstransfer(type = "binary", ...)
}
\arguments{
\item{type}{character, indicating the type of package to install. "binary" is
the default.}

\item{...}{other inputs passed to install.pacakges.}
}
\value{
Invisible 'NULL'.
}
\description{
Install gamstransfer package
}
