% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_gdxrrw.R
\name{install_gdxrrw}
\alias{install_gdxrrw}
\title{Install gdxrrw package}
\usage{
install_gdxrrw(version = "latest", ...)
}
\arguments{
\item{version}{character string defining the version of the package to
install. Default to "latest" for the latest release. Could also be "dev"
for the development version or a commit, tag, or branch name from the gdxrrw
repo (<https://github.com/GAMS-dev/gdxrrw>).}

\item{...}{other inputs passed to install_github.}
}
\value{
Invisible 'NULL'.
}
\description{
Install gdxrrw package
}
