% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{mock_transfer_adapter}
\alias{mock_transfer_adapter}
\title{Create a mock transfer adapter}
\usage{
mock_transfer_adapter()
}
\value{
A \code{gams_transfer_adapter} object.
}
\description{
The mock adapter is intended for unit tests. It does not create a valid GDX
file.
}
