% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{model_ir}
\alias{model_ir}
\title{Build a normalized model IR}
\usage{
model_ir(problem)
}
\arguments{
\item{problem}{A problem created by \code{\link[=gams_problem]{gams_problem()}}.}
}
\value{
A list with class \code{gams_ir}.
}
\description{
Build a normalized model IR
}
