% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{model_status}
\alias{model_status}
\alias{solver_status}
\title{Inspect solve status}
\usage{
model_status(result)

solver_status(result)
}
\arguments{
\item{result}{A result created by \code{solve()}.}
}
\value{
A \code{gams_status_value} object.
}
\description{
Inspect solve status
}
