% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{objective_value}
\alias{objective_value}
\title{Extract objective value}
\usage{
objective_value(result)
}
\arguments{
\item{result}{A result created by \code{solve()}.}
}
\value{
A numeric scalar.
}
\description{
Extract objective value
}
