% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gdx.R
\name{prepare_map4gdx}
\alias{prepare_map4gdx}
\title{Prepare mapping for export to gdx}
\usage{
prepare_map4gdx(x, symName, domains, symText, na.rm = TRUE)
}
\arguments{
\item{x}{R variable to be prepared for exportation.}

\item{symName}{Symbol name to be used in gdx file. Use by default the name
of the variable to be exported.}

\item{domains}{optional vector of characters defining the domain over which
the parameter is defined. Default to "*".}

\item{symText}{optional character string defining the explanatory text of
the variable. Default to "".}
}
\value{
A 'data.frame' that can be used in the function 'write_gdx'.
}
\description{
Prepare mapping for export to gdx
}
