% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gdx.R
\name{read_gdx}
\alias{read_gdx}
\title{Read gdx files}
\usage{
read_gdx(
  files,
  symName,
  col_names = NULL,
  names = NULL,
  attribute = "l",
  data_type = "tb",
  ...
)
}
\arguments{
\item{files}{path to one or several gdx files}

\item{symName}{symbol name to read in the gdx}

\item{col_names}{a vector of optional names for the columns. The default is
to use the names in the gdx if existing.}

\item{names}{a vector of optional names in case several gdx files are
imported. The default is to use the gdx file name.}

\item{attribute}{character string: the attribute for variables. Possible
values are '"l"' (the default) specifies the level, '"m"' specificies the
marginal, '"lo"' specificies the lower bound, and '"up"' specifies the upper
bound.}

\item{data_type}{character string: the type of data to output. Possible
values are '"tb"' (the default) for a tibble, '"dt"' for a data.table, and
'"df"' for a data.frame.}

\item{...}{additional arguments to be passed to gdxrrw functions.}
}
\value{
A 'tibble()', a 'data.table', or a 'data.frame'.
}
\description{
Read gdx files in a tidy way
}
\examples{
fpath <- system.file("extdata", "trnsport.gdx", package="gdxrrw")
read_gdx(fpath, "a")
read_gdx(fpath, "x")
read_gdx(fpath, "i", te = FALSE)
}
