% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_info_gdxfile.R
\name{read_info_gdxfile}
\alias{read_info_gdxfile}
\title{Function to read and save the information on all the sets, scalars, parameters and variables from a gdx file. Function used in import_gdxfile()}
\usage{
read_info_gdxfile(filename)
}
\arguments{
\item{filename}{Name of the gdxfile to be imported.}
}
\value{
Dataframe "parinfo" with info on all sets, scalars, parameters, and variables in the gdxfile
}
\description{
Function to read and save the information on all the sets, scalars, parameters and variables from a gdx file. Function used in import_gdxfile()
}
\examples{
\dontrun{read_info_gdxfile("trnsport.gdx")}
}
