% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{read_solution_gdx}
\alias{read_solution_gdx}
\title{Read a solved result GDX file}
\usage{
read_solution_gdx(problem, file)
}
\arguments{
\item{problem}{A problem created by \code{\link[=gams_problem]{gams_problem()}}.}

\item{file}{Result \code{.gdx} file containing GAMSr status scalars and selected
variable/equation records.}
}
\value{
A \code{gams_result_data} object.
}
\description{
Read a solved result GDX file
}
