% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgdx.var.R
\name{rgdx.var}
\alias{rgdx.var}
\title{Auxiliary function to read a variable from a gdx file (like rgdx.set). Function used in import_gdxfile().}
\usage{
rgdx.var(varname)
}
\arguments{
\item{varname}{Name of the gdx variable to be imported}
}
\value{
object with all the information from the variable
}
\description{
Auxiliary function to read a variable from a gdx file (like rgdx.set). Function used in import_gdxfile().
}
\examples{
\dontrun{rgdx.var("Z")}
}
