% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgdx_var.R
\name{rgdx_var}
\alias{rgdx_var}
\title{Auxiliary function to read a variable from a gdx file (like rgdx.set). Function used in import_gdxfile().}
\usage{
rgdx_var(varname)
}
\arguments{
\item{varname}{Name of the gdx variable to be imported}
}
\value{
object with all the information from the variable
}
\description{
Auxiliary function to read a variable from a gdx file (like rgdx.set). Function used in import_gdxfile().
}
\examples{
\dontrun{rgdx_var("Z")}
}
