% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{set_fixed}
\alias{set_fixed}
\title{Fix variable values}
\usage{
set_fixed(variable, records)
}
\arguments{
\item{variable}{A variable created by \code{\link[=gams_variable]{gams_variable()}}.}

\item{records}{Bound records in the same shape accepted by \code{\link[=gams_parameter]{gams_parameter()}}.}
}
\value{
A copy of \code{variable} with fixed-value records attached.
}
\description{
Fix variable values
}
