% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-functions.R
\name{set_functions}
\alias{set_functions}
\alias{gams_ord}
\alias{gams_card}
\title{Ordered-set position and cardinality}
\usage{
gams_ord(set)

gams_card(set)
}
\arguments{
\item{set}{A set created by \code{\link[=gams_set]{gams_set()}} or \code{\link[=gams_alias]{gams_alias()}}.}
}
\value{
A symbolic GAMSr expression.
}
\description{
\code{gams_ord()} returns the one-based position of a set element in GAMS's
internal order. \code{gams_card()} returns the number of elements in a set.
}
