% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{set_level}
\alias{set_level}
\title{Set variable starting levels}
\usage{
set_level(variable, records)
}
\arguments{
\item{variable}{A variable created by \code{\link[=gams_variable]{gams_variable()}}.}

\item{records}{Bound records in the same shape accepted by \code{\link[=gams_parameter]{gams_parameter()}}.}
}
\value{
A copy of \code{variable} with level records attached.
}
\description{
Set variable starting levels
}
