% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{set_upper_bound}
\alias{set_upper_bound}
\title{Set variable upper bounds}
\usage{
set_upper_bound(variable, records)
}
\arguments{
\item{variable}{A variable created by \code{\link[=gams_variable]{gams_variable()}}.}

\item{records}{Bound records in the same shape accepted by \code{\link[=gams_parameter]{gams_parameter()}}.}
}
\value{
A copy of \code{variable} with upper-bound records attached.
}
\description{
Set variable upper bounds
}
