% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{transfer_symbols}
\alias{transfer_symbols}
\title{Extract canonical transfer symbols}
\usage{
transfer_symbols(problem)
}
\arguments{
\item{problem}{A problem created by \code{\link[=gams_problem]{gams_problem()}}, or a model created by
\code{\link[=gams_model]{gams_model()}}.}
}
\value{
A \code{gams_transfer_symbols} object containing sets and parameters in
canonical long-form records.
}
\description{
Extract canonical transfer symbols
}
