% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-attribute.R
\name{variable_attribute_expressions}
\alias{variable_attribute_expressions}
\alias{gams_level}
\alias{gams_marginal}
\alias{gams_lower_bound}
\alias{gams_upper_bound}
\title{Symbolic variable attribute expressions}
\usage{
gams_level(variable)

gams_marginal(variable)

gams_lower_bound(variable)

gams_upper_bound(variable)
}
\arguments{
\item{variable}{A scalar variable or an indexed variable reference.}
}
\value{
A symbolic variable-attribute expression.
}
\description{
These helpers create references to GAMS variable attributes for use in
assignments and dollar conditions.
}
