% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{variable_values}
\alias{variable_values}
\title{Extract variable records}
\usage{
variable_values(result, variable)
}
\arguments{
\item{result}{A result created by \code{solve()}.}

\item{variable}{A variable object or variable name.}
}
\value{
A data frame of variable records.
}
\description{
Extract variable records
}
