% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_gamstransfer.R
\name{where_is_gams}
\alias{where_is_gams}
\title{Find the location of GAMS}
\usage{
where_is_gams()
}
\value{
A 'character'.
}
\description{
Three strategies are tried in the following order:
\enumerate{
\item Find GAMS in the path.
\item Find GAMS from the environment variable GAMSDIR.
\item Find GAMS in the standard installation folder.
}
}
