% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gdx.R
\name{write_gdx}
\alias{write_gdx}
\title{Write gdx files}
\usage{
write_gdx(gdxName, ...)
}
\arguments{
\item{gdxName}{path to one gdx file}

\item{...}{additional arguments to be passed to the function wgdx.lst.}
}
\description{
Write gdx files
}
\examples{
a <- data.frame(i = c("a", "b", "c"), val = 1:3)
a <- prepare_par4gdx(a)
i <- a[,"i", drop = FALSE]
i <- prepare_set4gdx(i)
gdxfile <- tempfile(fileext = ".gdx")
write_gdx(gdxfile, a, i)
}
