% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{write_input_gdx}
\alias{write_input_gdx}
\title{Write input data to a GDX file}
\usage{
write_input_gdx(
  problem,
  file,
  adapter = gamstransfer_adapter(),
  overwrite = FALSE
)
}
\arguments{
\item{problem}{A problem created by \code{\link[=gams_problem]{gams_problem()}}, or a model created by
\code{\link[=gams_model]{gams_model()}}.}

\item{file}{Output \code{.gdx} file path.}

\item{adapter}{Transfer adapter. Defaults to \code{\link[=gamstransfer_adapter]{gamstransfer_adapter()}}.}

\item{overwrite}{Whether to overwrite an existing file.}
}
\value{
A \code{gams_transfer_write} object.
}
\description{
Write input data to a GDX file
}
