% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gams.R
\name{gams}
\alias{gams}
\title{Launch GAMS from command line}
\usage{
gams(gmsfile, options = list(), save, restart, gdx, envvar, wd, ...)
}
\arguments{
\item{gmsfile}{path to one gms file}

\item{options}{list of options}

\item{save}{path to restart file}

\item{restart}{path to restart file}

\item{gdx}{path to gdx file}

\item{wd}{working directory from which to launch the file (default to current
working directory)}

\item{envar}{list of environmental variables}
}
\value{
A 'tibble()', a 'data.table', or a 'data.frame'.
}
\description{
Launch GAMS from command line
}
\examples{
fpath <- file.path(sub(";.*$", "", Sys.getenv("GAMSDIR")),
                   "gamslib_ml", "trnsport.1")
gams(fpath, options = list(output = "NUL", lp = "cplex"))
}
