% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gams.R
\name{gams}
\alias{gams}
\title{Launch GAMS from command line}
\usage{
gams(gmsfile, options = list(), save, restart, gdx, envvar, wd, ...)
}
\arguments{
\item{gmsfile}{path to one gms file}

\item{options}{list of options}

\item{save}{path to restart file}

\item{restart}{path to restart file}

\item{gdx}{path to gdx file}

\item{envvar}{list of environmental variables}

\item{wd}{working directory from which to launch the file (default to current
working directory)}

\item{...}{further argument to be passed to 'system'}
}
\value{
A 'tibble()' with three columns: 'gmsfile' containing the name of the
gms file that was launched, 'return_code' containing the return code
following the call, and 'status' a list column containing a 'tibble()' of
solver and model status. The most standard return codes are
\itemize{
\item 0 for normal return
\item 2 for compilation error
\item 3 for execution error
}

see \url{https://www.gams.com/latest/docs/UG_GAMSReturnCodes.html#UG_GAMSReturnCodes_ListOfErrorCodes}
for the full list of GAMS return codes.
For solver status codes, see \url{https://www.gams.com/latest/docs/UG_GAMSOutput.html#UG_GAMSOutput_SolverStatus}
and for model status, see \url{https://www.gams.com/latest/docs/UG_GAMSOutput.html#UG_GAMSOutput_ModelStatus}.
}
\description{
Launch GAMS from command line
}
\examples{
fpath <- file.path(sub(";.*$", "", Sys.getenv("GAMSDIR")),
                   "gamslib_ml", "trnsport.1")
gams(fpath, options = list(output = "NUL", lp = "cplex"))
}
\seealso{
\code{\link[=system]{system()}} for the list of options of the command used to launch
GAMS.
}
