% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.R
\name{gams_sum}
\alias{gams_sum}
\title{Sum a symbolic expression over one or more sets}
\usage{
gams_sum(index, expression)
}
\arguments{
\item{index}{A GAMSr set or a set vector created with \code{c(i, j)}.}

\item{expression}{A GAMSr symbolic expression.}
}
\value{
A \code{gams_expr_sum} object.
}
\description{
Sum a symbolic expression over one or more sets
}
