% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_gdxrrw.R
\name{install_gdxrrw}
\alias{install_gdxrrw}
\title{Install gdxrrw package}
\usage{
install_gdxrrw(version = "latest", type = "platform_default")
}
\arguments{
\item{version}{character string defining the version of the package to
install. Default to "latest".}

\item{type}{character, indicating the type of package to download and
install. Possible values are '"platform_default"' (the default) and
'"source"'. Except on Windows, the platform default is to install from
source. Any other value will be interpreted as installation from binary.}
}
\value{
Invisible 'NULL'.
}
\description{
Install gdxrrw package
}
