% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gdx.R
\name{read_gdx}
\alias{read_gdx}
\title{Read gdx files}
\usage{
read_gdx(
  files,
  symName = NULL,
  col_names = NULL,
  attributes = "l",
  data_type = "tb",
  factors_as_strings = TRUE,
  names = NULL,
  names_to = "name"
)
}
\arguments{
\item{files}{path to one or several gdx files}

\item{symName}{symbol name to read in the gdx. If NULL (default) returns a
data.frame of the list of symbols in the gdx.}

\item{col_names}{a vector of optional names for the columns.}

\item{attributes}{character vector: the attributes to keep for variables,
equations and sets. Possible values are '"l"' (the default) specifies the
level, '"m"' specificies the marginal, '"lo"' specificies the lower bound,
'"up"' specifies the upper bound, '"s"' specificies the scale, and '"te"'
specified the text for sets.}

\item{data_type}{character string: the type of data to output. Possible
values are '"tb"' (the default) for a tibble, '"dt"' for a data.table, and
'"df"' for a data.frame.}

\item{names}{a vector of optional names in case several gdx files are
imported. The default is to use the gdx file name.}

\item{names_to}{character string specifying the new column to create to
store the file names.}

\item{factors_as_string}{logical (default is TRUE) specifying whether factors
should be transformed in strings.}
}
\value{
A 'tibble()', a 'data.table', or a 'data.frame'.
}
\description{
Read gdx files in a tidy way
}
\examples{
fpath <- system.file("extdata", "trnsport.gdx", package = "gamsr")
read_gdx(fpath, "a")
read_gdx(fpath)
read_gdx(fpath, "x")
read_gdx(fpath, "f")
read_gdx(fpath, "i", attributes = "te")
}
