% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_gamstransfer.R
\name{where_is_gams}
\alias{where_is_gams}
\title{Find the location of GAMS}
\usage{
where_is_gams()
}
\value{
A 'character'.
}
\description{
Three strategies are tried in the following order:
1. Find GAMS in the path.
2. Find GAMS from the environment variable GAMSDIR.
3. Find GAMS in the standard installation folder.
}
